+++ .github/workflows/rust.yaml
@@ -26,5 +26,7 @@ jobs:
         run: cargo build --verbose
       - name: Clippy
         run: cargo clippy --verbose -- -D warnings
-      - name: Run tests
+      - name: Run Async Tests
         run: cargo test --verbose
+      - name: Run Blocking Tests
+        run: cargo test --verbose --features blocking
