+++ src/prelude.rs
@@ -1,0 +1,2 @@
+pub use crate::client::Client as MsucClient;
+pub use crate::client::SearchResultsStreamer;
